installed django, added rough cost calculation
ive installed django to learn about/test out the app with databases.
the cost calculation is a work in progress, im planning on using a gcode converter to approximate the etching laser path.

so we'd take the svg/dxf file and convert it to gcode and write a script that can get the length of the path the laser has to travel directly from the gcode.